requirements.txt
torch
torchvision
numpy
opencv-python-headless
PyYAML
Pillow
transformers
timm
packaging